mounting work started
