+++ .github/FUNDING.yml
@@ -1,2 +1,1 @@
 patreon: user?u=32109749
-custom: ["paypal.me/abhinabaghosh"]
